+++ src/api/apis/category.ts
@@ -1,0 +1,5 @@
+import { api } from '@api/instance';
+
+export const CategoryAPI = {
+  getCategoryList: () => api.get(`/api/v1/category`),
+};
