+++ src/api/apis/mentor.ts
@@ -1,0 +1,7 @@
+import { api } from '@api/instance';
+import { IApplyMentor } from 'types/api';
+
+export const MentorAPI = {
+  applyMentor: (payload: IApplyMentor) =>
+    api.post('/api/v1/mentor/apply', { ...payload }),
+};
