+++ src/api/apis/select.ts
@@ -1,0 +1,5 @@
+import { api } from '@api/instance';
+
+export const SelectAPI = {
+  selectCategory: () => api.get('api/v1/category'),
+};
