+++ src/components/domains/createStudy/Buttons/index.tsx
@@ -1,11 +1,11 @@
 import { AxiosResponse } from 'axios';
 import Swal from 'sweetalert2';
 
-import { PostAPI } from '@api/api';
 import { PostRequestDto } from 'types/api.types';
 import { useCustomNavigate } from '@hooks/useCustomNavigate';
 import { useAddPost, useUpdatePost } from '@hooks/query/usePostQuery';
 import { validateFormData } from '@utils/validateData';
+import { PostAPI } from '@api/apis/post';
 
 interface SubmitButtonProps {
   className: string;
